+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/ast/GenericToken.java
@@ -56,8 +56,13 @@ public interface GenericToken {
 
     /**
      * Gets a unique integer representing the kind of token this is.
-     *
      * The semantics of this kind depend on the language.
+     *
+     * <p><strong>Note:</strong> This is an experimental API.
+     *
+     * <p>The returned constants can be looked up in the language's "*ParserConstants",
+     * e.g. CppParserConstants or JavaParserConstants. These constants are considered
+     * internal API and may change at any time when the language's grammar is changed.
      */
     @Experimental
     int getKind();